src/test/kotlin/org/carth/aoc21/day17/Day17Test.kt
package org.carth.aoc21.day17

import org.carth.aoc21.common.Resources.inputAsListOfString
import org.carth.aoc21.common.Resources.testInputAsListOfString
import org.junit.jupiter.api.Assertions.assertEquals
import org.junit.jupiter.api.MethodOrderer
import org.junit.jupiter.api.Nested
import org.junit.jupiter.api.Order
import org.junit.jupiter.api.Test
import org.junit.jupiter.api.TestMethodOrder

private const val DAY = "day17"

class Day17Tests {

    @Nested
    @TestMethodOrder(MethodOrderer.OrderAnnotation::class)
    inner class Day17PartOneTest {

        @Test
        @Order(1)
        fun `solve part one of sample`() {
            // Act
            val answer = Day17(testInputAsListOfString(DAY)).solvePartOne()

            // Assert
            assertEquals(45, answer)
        }

        @Test
        @Order(2)
        fun `solve part one`() {
            // Act
            val answer = Day17(inputAsListOfString(DAY)).solvePartOne()

            // Assert
            assertEquals(4560, answer)
        }
    }

    @Nested
    @TestMethodOrder(MethodOrderer.OrderAnnotation::class)
    inner class Day17PartTwoTest {

        @Test
        @Order(1)
        fun `solve part two of sample`() {
            // Act
            val answer = Day17(testInputAsListOfString(DAY)).solvePartTwo()

            // Assert
            assertEquals(112, answer)
        }

        @Test
        @Order(2)
        fun `solve part two`() {
            // Act
            val answer = Day17(inputAsListOfString(DAY)).solvePartTwo()
            // Assert
            assertEquals(3344, answer)
        }
    }
}